Add ESLint and api-types file
