+++ client-js/package.json
@@ -17,6 +17,9 @@
   "license": "ISC",
   "devDependencies": {
     "@rollup/plugin-typescript": "^8.3.4",
+    "@typescript-eslint/eslint-plugin": "^5.34.0",
+    "@typescript-eslint/parser": "^5.34.0",
+    "eslint": "^8.22.0",
     "rollup": "^2.78.0",
     "rollup-plugin-dts": "^4.2.2",
     "tslib": "^2.4.0",
