+++ client-js/src/clients/RecordsClient.ts
@@ -1,15 +1,5 @@
-import { Country, Metric, Record, Period, Player, PlayerBuild, PlayerType } from '../../../server/src/utils';
+import { RecordLeaderboardFilter, GetRecordLeaderboardResponse } from '../api-types';
 import { sendGetRequest } from '../utils';
-
-export interface RecordLeaderboardFilter {
-  metric: Metric;
-  period: Period | string;
-  country?: Country;
-  playerType?: PlayerType;
-  playerBuild?: PlayerBuild;
-}
-
-export type GetRecordLeaderboardResponse = Array<Record & { player: Player }>;
 
 export default class RecordsClient {
   getRecordLeaderboard(filter: RecordLeaderboardFilter) {
